added list students template and refactor
