+++ src/mooc/test/tv/codely/mooc/students_register/application/incrementer/IncrementStudentsCounterOnStudentRegisteredShould.java
@@ -1,4 +1,4 @@
-package tv.codely.mooc.students_register.application;
+package tv.codely.mooc.students_register.application.incrementer;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -6,6 +6,8 @@ import tv.codely.mooc.students.domain.StudentId;
 import tv.codely.mooc.students.domain.StudentIdMother;
 import tv.codely.mooc.students.domain.StudentRegisteredDomainEventMother;
 import tv.codely.mooc.students_register.StudentsRegisterModuleUnitTestCase;
+import tv.codely.mooc.students_register.application.incrementer.IncrementerStudentsOnStudentRegistered;
+import tv.codely.mooc.students_register.application.incrementer.StudentsRegisterIncrementer;
 import tv.codely.mooc.students_register.domain.StudentsRegister;
 import tv.codely.mooc.students_register.domain.StudentsRegisterMother;
 import tv.codely.shared.domain.student.StudentRegisteredDomainEvent;
